Install openstack-health in editable mode
Due to the magic of non-determinism that is python packaging for
whatever reason the wsgi_script entrypoint is not being generated
when installing openstack-health in non-editable mode. So while we
wait for the local witch doctor to remove the curse from the system
this switches to using editable installs to move forward.

As part of this change the exec for running pip to install
openstack-health is also improved.

Change-Id: If5e166d6cba0553f160eabde491d0230e4834f01