Only run mv if we generated a new js build
This commit adds a missing refreshonly from the exec to run the mv
which moves the build dir to the output dir. Previously we were
running this on every exec which caused failures when a new build
wasn't run. This should address this by only running the mv when we
build the js.

Change-Id: Id9d48c0e5da4567b0afefad22513c391b95e9bc0