Add E-R repo and revision to class
The openstack_health class has been recently extended to include
deployment to E-R, a new data source for OpenStack Health.
The E-R repo and revions are hardcoded in the class, which prevents
using the new E-R feature in downstream deployments that have an own
E-R repo.

Adding repo and revision at class level solves that.

Change-Id: Ifde71192f0cc618480be95feb77e712f8c27356b